Visit Logger/requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-webpack-loader==1.6.0
djangorestframework==3.13.1
gunicorn==20.1.0
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
psycopg2==2.8.4
psycopg2-binary